added Sinai billiard
