+++ src/utils/billiardutils.jl
@@ -1,13 +1,3 @@
-#=
-struct Billiard  <: AbsBilliard
-    boundary :: Vector{<:AbsCurve}
-    area :: Float64
-    length:: Float64
-    domain:: Function
-end
-=#
-
-#include("../abstracttypes.jl")
 
 function real_length(billiard::Bi) where Bi<:AbsBilliard
     L = 0.0
@@ -57,16 +47,6 @@ function is_inside(billiard::Bi, pts::AbstractArray) where Bi<:AbsBilliard
     end 
 end
 
-function is_inside(billiard::Bi, x_grid::AbstractArray, y_grid::AbstractArray) where Bi<:AbsBilliard
-    let curves = billiard.boundary
-        inside = is_inside(curves[1], x_grid, y_grid)
-        for i in 2:length(curves)
-            inside = inside .& is_inside(curves[i], x_grid, y_grid)
-        end
-        return inside
-    end 
-end
-
 
 function boundary_limits(curves; grd=1000, type=Float64) 
     x_bnd = Vector{Any}()
@@ -87,3 +67,15 @@ function boundary_limits(curves; grd=1000, type=Float64)
     #dy =  ylim[2] - ylim[1]
     return xlim, ylim #,dx,dy
 end
+
+#=
+function is_inside(billiard::Bi, x_grid::AbstractArray, y_grid::AbstractArray) where Bi<:AbsBilliard
+    let curves = billiard.boundary
+        inside = is_inside(curves[1], x_grid, y_grid)
+        for i in 2:length(curves)
+            inside = inside .& is_inside(curves[i], x_grid, y_grid)
+        end
+        return inside
+    end 
+end
+=#
